# src/api-contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DailyNagger.Server | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5010/"
    }
  ],
  "paths": {
...
    "/api/todays-nag-plan": {
      "get": {
        "tags": [
          "NagPlan"
        ],
        "parameters": [
          {
            "name": "communityId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NagPlanDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/nag-plan": {
      "get": {
        "tags": [
          "NagPlan"
        ],
        "parameters": [
          {
            "name": "communityId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NagPlanDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NaggerPinnedByDto": {
        "enum": [
          "None",
          "User",
          "Llm",
          "Community"
        ]
      },
      "NagPlanDto": {
        "required": [
          "date",
          "nags"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "nags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NagPlanNaggerDto"
            }
          }
        }
      },
      "NagPlanNaggerDto": {
        "required": [
          "id",
          "title",
          "activeLogDueOn",
          "expiresOn",
          "targetTime",
          "isDeactivated",
          "pinnedBy",
          "updatedAt",
          "updatedByClientId",
          "updatedByDeviceName",
          "updatedByDeviceModel",
          "scheduleRules",
          "taskLog",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "activeLogDueOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "expiresOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "targetTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "isDeactivated": {
            "type": "boolean"
          },
          "pinnedBy": {
            "$ref": "#/components/schemas/NaggerPinnedByDto"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedByClientId": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedByDeviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedByDeviceModel": {
            "type": [
              "null",
              "string"
            ]
          },
          "scheduleRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleRuleDto"
            }
          },
          "taskLog": {
            "$ref": "#/components/schemas/TaskLogDto"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ScheduleRuleDto": {
        "required": [
          "id",
          "ruleType",
          "ruleJson"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ruleType": {
            "$ref": "#/components/schemas/ScheduleRuleTypeDto"
          },
          "ruleJson": {
            "type": "string"
          }
        }
      },
      "ScheduleRuleTypeDto": {
        "enum": [
          "Weekday",
          "Date",
          "Holiday"
        ]
      },
...
      "TaskEntryDto": {
        "required": [
          "id",
          "taskLogId",
          "parentTaskItemId",
          "label",
          "description",
          "valueType",
          "tag",
          "value",
          "lastTaskRunReferenceValue",
          "rolloverBehavior",
          "interactionAt",
          "interactionTimeZone",
          "interactionLocale",
          "interactionMood",
          "interactionMoodAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taskLogId": {
            "type": "string",
            "format": "uuid"
          },
          "parentTaskItemId": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "valueType": {
            "$ref": "#/components/schemas/TaskEntryValueTypeDto"
          },
          "tag": {
            "type": [
              "null",
              "string"
            ]
          },
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastTaskRunReferenceValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "rolloverBehavior": {
            "$ref": "#/components/schemas/TaskEntryRolloverBehaviorDto"
          },
          "interactionAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "interactionTimeZone": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionLocale": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMood": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMoodAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "TaskEntryRolloverBehaviorDto": {
        "enum": [
          "Remove",
          "MoveValueToHistory",
          "CarryOverValue"
        ]
      },
      "TaskEntryValueTypeDto": {
        "enum": [
          "Text",
          "Integer",
          "Decimal",
          "Boolean"
        ]
      },
...
      "TaskItemDto": {
        "required": [
          "id",
          "taskLogId",
          "parentTaskItemId",
          "name",
          "taskEntries",
          "taskItems",
          "tag",
          "isDone",
          "rolloverBehavior",
          "interactionAt",
          "interactionTimeZone",
          "interactionLocale",
          "interactionMood",
          "interactionMoodAt",
          "descendantTaskItemCount",
          "doneDescendantTaskItemCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taskLogId": {
            "type": "string",
            "format": "uuid"
          },
          "parentTaskItemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "taskEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskEntryDto"
            }
          },
          "taskItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskItemDto"
            }
          },
          "tag": {
            "type": [
              "null",
              "string"
            ]
          },
          "isDone": {
            "type": "boolean"
          },
          "rolloverBehavior": {
            "$ref": "#/components/schemas/TaskItemRolloverBehaviorDto"
          },
          "interactionAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "interactionTimeZone": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionLocale": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMood": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMoodAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "descendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          },
          "doneDescendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TaskItemRolloverBehaviorDto": {
        "enum": [
          "Keep",
          "RemoveWhenDone"
        ]
      },
      "TaskLogDto": {
        "required": [
          "id",
          "nagId",
          "copiedFromTaskLogId",
          "closedOn",
          "updatedAt",
          "updatedByClientId",
          "updatedByDeviceName",
          "updatedByDeviceModel",
          "version",
          "taskItems",
          "tag",
          "descendantTaskItemCount",
          "doneDescendantTaskItemCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nagId": {
            "type": "string",
            "format": "uuid"
          },
          "copiedFromTaskLogId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "closedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedByClientId": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedByDeviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedByDeviceModel": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "taskItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskItemDto"
            }
          },
          "tag": {
            "type": [
              "null",
              "string"
            ]
          },
          "descendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          },
          "doneDescendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}